# Generated with AI Add verbose mode to evaluate_recommender_algorithm function
Modified the evaluate_recommender_algorithm function to include a verbose mode that shows a progress bar when the function is running. This improves readability and user experience when processing bigger datasets which take a considerable time to analyze. The verbose mode can be turned on/off using the 'batch_verbose' setting in the configuration file. This setting has now been added to the default configuration parameters.